+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### documentdb v0.102-0 (Unreleased) ###
 * Support index pushdown for vector search queries *[Bugfix]*
+* Support exact search for vector search queries *[Feature]*
 * Inline $match with let in $lookup pipelines as JOIN Filter *[Perf]*
 
 ### documentdb v0.101-0 (February 12, 2025) ###